REFACTOR-001: extract ProjectLoadStates and useEditorShortcuts from EditorView
Pure structural move, no behavior change.

- ProjectLoadStates owns the loading/notFound/error branches (the first
  three Match arms of the top-level Switch), preserving their mutual
  exclusivity with its own internal Switch/Match.
- useEditorShortcuts is a plain function of the same dependencies
  EditorView already held (audio controls, session, the piano roll's
  lifted state, guide/pack-browser signals) that installs the KEY-01
  shortcut mapping. A true panel-contributed registration seam (per the
  issue's suggestion) would need a real architecture change with only
  one consumer today; this still gets a future panel's shortcut needs
  into one small file instead of EditorView.tsx itself.

EditorView.tsx is now 427 lines (down from 750 on main), all of it
composition plus derived state and wiring that is genuinely shared
across the header, track editor, and shortcut layer.

EditorView.test.tsx (610 lines, unmodified) passes unchanged — the
evidence this refactor did not change behavior.

Closes #142
